Set default jointype for polygons-offsets to round.
Part of why support is such a mess, is that the default offsets when done with jtMitre where terrible: Not only is jtMitre (no matter wether its cheaper to calculate) very probably the wrong choice in the first place, but it is also wrong in the sense that it can generate asymmetrical polygons from polygons that where previously symmetrical (though you'd think that wouldn't be possible).

In any case, we're still investigating this: At the moment it's not clear if we really want to set the default to this in main for example.

part of spike: CURA-9996