Help support with horizontal offset applied be apart from the model.
Previously, we couldn't set the horizontal support expansion very high, as the expansion wraps around the model. This can be counteracted by not allowing the _expansion_ of support near the model, but still allow (previously calculated, nonexpanded) support as close as the x/y(/z) support-offset allows.

part of investigative spike CURA-9996